Used shared storage_options utility to pass credentials through
