+++ daft/delta_lake/delta_lake_scan.py
@@ -3,24 +3,18 @@ from __future__ import annotations
 import logging
 import os
 from collections.abc import Iterator
-from typing import Any
-from urllib.parse import urlparse
 
 from deltalake.table import DeltaTable
 
 import daft
 from daft.daft import (
-    AzureConfig,
     FileFormatConfig,
-    GCSConfig,
-    IOConfig,
-    NativeStorageConfig,
     ParquetSourceConfig,
     Pushdowns,
-    S3Config,
     ScanTask,
     StorageConfig,
 )
+from daft.io.object_store_options import storage_config_to_storage_options
 from daft.io.scan import PartitionField, ScanOperator
 from daft.logical.schema import Schema
 
@@ -30,7 +24,7 @@ logger = logging.getLogger(__name__)
 class DeltaLakeScanOperator(ScanOperator):
     def __init__(self, table_uri: str, storage_config: StorageConfig) -> None:
         super().__init__()
-        storage_options = _storage_config_to_storage_options(storage_config, table_uri)
+        storage_options = storage_config_to_storage_options(storage_config, table_uri)
         self._table = DeltaTable(table_uri, storage_options=storage_options)
         self._storage_config = storage_config
         self._schema = Schema.from_pyarrow_schema(self._table.schema().to_pyarrow())
@@ -165,66 +159,3 @@ class DeltaLakeScanOperator(ScanOperator):
 
     def can_absorb_select(self) -> bool:
         return True
-
-
-def _storage_config_to_storage_options(storage_config: StorageConfig, table_uri: str) -> dict[str, str] | None:
-    """
-    Converts the Daft storage config to a storage options dict that deltalake/object_store
-    understands.
-    """
-    config = storage_config.config
-    assert isinstance(config, NativeStorageConfig)
-    io_config = config.io_config
-    return _io_config_to_storage_options(io_config, table_uri)
-
-
-def _io_config_to_storage_options(io_config: IOConfig, table_uri: str) -> dict[str, str] | None:
-    scheme = urlparse(table_uri).scheme
-    if scheme == "s3" or scheme == "s3a":
-        return _s3_config_to_storage_options(io_config.s3)
-    elif scheme == "gcs" or scheme == "gs":
-        return _gcs_config_to_storage_options(io_config.gcs)
-    elif scheme == "az" or scheme == "abfs":
-        return _azure_config_to_storage_options(io_config.azure)
-    else:
-        return None
-
-
-def _s3_config_to_storage_options(s3_config: S3Config) -> dict[str, str]:
-    storage_options: dict[str, Any] = {}
-    if s3_config.region_name is not None:
-        storage_options["region"] = s3_config.region_name
-    if s3_config.endpoint_url is not None:
-        storage_options["endpoint_url"] = s3_config.endpoint_url
-    if s3_config.key_id is not None:
-        storage_options["access_key_id"] = s3_config.key_id
-    if s3_config.session_token is not None:
-        storage_options["session_token"] = s3_config.session_token
-    if s3_config.access_key is not None:
-        storage_options["secret_access_key"] = s3_config.access_key
-    if s3_config.use_ssl is not None:
-        storage_options["allow_http"] = "false" if s3_config.use_ssl else "true"
-    if s3_config.verify_ssl is not None:
-        storage_options["allow_invalid_certificates"] = "false" if s3_config.verify_ssl else "true"
-    if s3_config.connect_timeout_ms is not None:
-        storage_options["connect_timeout"] = str(s3_config.connect_timeout_ms) + "ms"
-    if s3_config.anonymous:
-        storage_options["skip_signature"] = "true"
-    return storage_options
-
-
-def _azure_config_to_storage_options(azure_config: AzureConfig) -> dict[str, str]:
-    storage_options = {}
-    if azure_config.storage_account is not None:
-        storage_options["account_name"] = azure_config.storage_account
-    if azure_config.access_key is not None:
-        storage_options["access_key"] = azure_config.access_key
-    if azure_config.endpoint_url is not None:
-        storage_options["endpoint"] = azure_config.endpoint_url
-    if azure_config.use_ssl is not None:
-        storage_options["allow_http"] = "false" if azure_config.use_ssl else "true"
-    return storage_options
-
-
-def _gcs_config_to_storage_options(_: GCSConfig) -> dict[str, str]:
-    return {}
